fix: fixes discrepancy in how SQL errors are managed
With the switch to the DBAL wrapper classes for PDO, SQL errors are
now thrown as PDOExceptions. This is the desired behavior, however
some functionality, specifically upgrade.php, relies on these
errors not being turned into exceptions. This commit provides a
middle-ground, where the exceptions are handled by CDash\Database
which we will eventually move away from.